+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * Implemented new Photometry class. It can extract from multiple reflected regions in multiprocessing and return numpy array instead of csv files.
 * The p-value analysis code has been refactored. The pipeline will apply both the photometry and model prediction starting from the DL3.  
 * Updated testing code. Load a named test-set composed by multiple timeseries.
+* Major refactoring of OnlinePhotometry.